refactor(storage): extract SQLiteFtsVecMixin from sqlite _base.py
Peel the 7 self-committing FTS5/sqlite-vec index-maintenance helpers
(_fts_upsert, _fts_delete, _fts_upsert_profile, _fts_delete_profile,
_vec_upsert, _vec_delete, _vec_knn_search) verbatim out of
SQLiteStorageBase into a new SQLiteFtsVecMixin (sqlite-only; no
supabase/postgres counterpart). Method bodies are byte-identical and
preserve each helper's internal self-commit.

Composed into SQLiteStorage ahead of SQLiteStorageBase; the residual
boot-time _migrate_vec_tables resolves self._vec_upsert through the MRO.
A type-only _vec_upsert annotation is declared on SQLiteStorageBase for
_migrate_vec_tables's benefit.
