+++ reflexio/server/services/storage/sqlite_storage/_base.py
@@ -582,6 +582,10 @@ class SQLiteStorageBase(RetentionMixin, BaseStorage):
     _delete_in_chunks: Any
     _delete_source_windows_for_user_playbook_ids: Any
 
+    # FTS/vec index helpers provided by SQLiteFtsVecMixin via the composed
+    # SQLiteStorage MRO; declared here for _migrate_vec_tables's benefit.
+    _vec_upsert: Any
+
     @staticmethod
     def handle_exceptions(func: Callable[..., Any]) -> Callable[..., Any]:
         @functools.wraps(func)
@@ -1519,113 +1523,6 @@ class SQLiteStorageBase(RetentionMixin, BaseStorage):
 
     def _current_timestamp(self) -> str:
         return datetime.now(UTC).isoformat()
-
-    # FTS helpers
-    def _fts_upsert(self, table: str, rowid: int, **text_fields: str | None) -> None:
-        """Insert or update an FTS row.  Deletes old entry first to avoid duplicates."""
-        with self._lock:
-            self.conn.execute(f"DELETE FROM {table} WHERE rowid = ?", (rowid,))
-            cols = list(text_fields.keys())
-            vals = [text_fields[c] or "" for c in cols]
-            placeholders = ",".join("?" for _ in cols)
-            col_str = ",".join(cols)
-            self.conn.execute(
-                f"INSERT INTO {table}(rowid, {col_str}) VALUES (?, {placeholders})",
-                [rowid, *vals],
-            )
-            self.conn.commit()
-
-    def _fts_delete(self, table: str, rowid: int) -> None:
-        with self._lock:
-            self.conn.execute(f"DELETE FROM {table} WHERE rowid = ?", (rowid,))
-            self.conn.commit()
-
-    def _fts_upsert_profile(self, profile_id: str, content: str) -> None:
-        """FTS for profiles uses profile_id TEXT as key column."""
-        with self._lock:
-            self.conn.execute(
-                "DELETE FROM profiles_fts WHERE profile_id = ?", (profile_id,)
-            )
-            self.conn.execute(
-                "INSERT INTO profiles_fts(profile_id, content) VALUES (?, ?)",
-                (profile_id, content),
-            )
-            self.conn.commit()
-
-    def _fts_delete_profile(self, profile_id: str) -> None:
-        with self._lock:
-            self.conn.execute(
-                "DELETE FROM profiles_fts WHERE profile_id = ?", (profile_id,)
-            )
-            self.conn.commit()
-
-    # Vec helpers (sqlite-vec)
-    def _vec_upsert(self, table: str, rowid: int, embedding: list[float]) -> None:
-        """Insert or update a vec table row. No-op when sqlite-vec is unavailable."""
-        if not self._has_sqlite_vec:
-            return
-        with self._lock:
-            self.conn.execute(f"DELETE FROM {table} WHERE rowid = ?", (rowid,))
-            self.conn.execute(
-                f"INSERT INTO {table}(rowid, embedding) VALUES (?, ?)",
-                (rowid, json.dumps(embedding)),
-            )
-            self.conn.commit()
-
-    def _vec_delete(self, table: str, rowid: int) -> None:
-        """Delete a vec table row. No-op when sqlite-vec is unavailable."""
-        if not self._has_sqlite_vec:
-            return
-        with self._lock:
-            self.conn.execute(f"DELETE FROM {table} WHERE rowid = ?", (rowid,))
-            self.conn.commit()
-
-    def _vec_knn_search(
-        self,
-        vec_table: str,
-        main_table: str,
-        query_embedding: list[float],
-        match_count: int,
-        conditions: list[str] | None = None,
-        params: list[Any] | None = None,
-    ) -> list[sqlite3.Row]:
-        """Run a native KNN search via sqlite-vec and join back to the main table.
-
-        Over-fetches from the KNN index (5x ``match_count``) so that post-filter
-        WHERE conditions (org, user, status, etc.) don't silently reduce the
-        result set below the requested count.
-
-        Args:
-            vec_table: Name of the vec0 virtual table.
-            main_table: Name of the main data table.
-            query_embedding: Query embedding vector.
-            match_count: Number of results to return.
-            conditions: Optional WHERE conditions for the main table.
-            params: Parameters for the conditions.
-
-        Returns:
-            Up to ``match_count`` rows from the main table, ordered by vector
-            distance (ascending).
-        """
-        knn_overfetch = match_count * 5
-        where_clause = " AND ".join(conditions) if conditions else "1=1"
-        sql = f"""SELECT m.* FROM {main_table} m
-                  JOIN (
-                      SELECT rowid, distance FROM {vec_table}
-                      WHERE embedding MATCH ?
-                      ORDER BY distance
-                      LIMIT ?
-                  ) v ON m.rowid = v.rowid
-                  WHERE {where_clause}
-                  ORDER BY v.distance
-                  LIMIT ?"""
-        all_params = [
-            json.dumps(query_embedding),
-            knn_overfetch,
-            *(params or []),
-            match_count,
-        ]
-        return self._fetchall(sql, all_params)
 
     # ------------------------------------------------------------------
     # Per-user data clear
