+++ reflexio/server/services/storage/sqlite_storage/base/__init__.py
@@ -1,5 +1,6 @@
 """SQLite storage base sub-mixins (peeled from _base.py)."""
 
 from ._deletion import SQLiteDeletionMixin
+from ._fts_vec import SQLiteFtsVecMixin
 
-__all__ = ["SQLiteDeletionMixin"]
+__all__ = ["SQLiteDeletionMixin", "SQLiteFtsVecMixin"]
